refactor(money): move the ledger into domain money
